Moved gRPC connection factory directory out of utils
