+++ src/main/java/emissary/grpc/pool/LoadBalancingPolicy.java
@@ -1,4 +1,4 @@
-package emissary.util.grpc.pool;
+package emissary.grpc.pool;
 
 import java.util.Locale;
